+++ README.md
@@ -33,3 +33,9 @@
     * Перевод под архитектуру MVVM (Продолжение "Чистая архетиктура")
 - 10-й Этап (Спринт 17)
     * Внедрение зависимостей (Koin)
+- 11-й Этап (Спринт 18)
+    * Добавлена "Медиатека" на Фрагментах
+- 12-й Этап (Спринт 19)
+    * Приложение переведено на SingleActivity
+    * Сделана навигация между Fragments
+    * Добавлен BottomNavigationView
