--- app/src/main/java/com/example/playlistmaker/library/ui/activity/LibraryActivity.kt
@@ -1,41 +1,0 @@
-package com.example.playlistmaker.library.ui.activity
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.R
-import com.example.playlistmaker.databinding.ActivityLibraryBinding
-import com.google.android.material.tabs.TabLayoutMediator
-
-class LibraryActivity : AppCompatActivity() {
-    private var _binding: ActivityLibraryBinding? = null
-    private val binding get() = _binding!!
-
-    private var _tabMediator: TabLayoutMediator? = null
-    private val tabMediator get() = _tabMediator!!
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        _binding = ActivityLibraryBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.viewPager.adapter = LibraryViewPageAdapter(
-            fragmentManager = supportFragmentManager,
-            lifecycle = lifecycle
-        )
-
-        binding.home.setOnClickListener { this.finish() }
-
-        _tabMediator = TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
-            when(position) {
-                0 -> tab.text = getString(R.string.favorite_tab)
-                1 -> tab.text = getString(R.string.playlist_tab)
-            }
-        }
-        tabMediator.attach()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        tabMediator.detach()
-    }
-}
