+++ app/src/main/java/com/example/playlistmaker/library/ui/fragments/LibraryFragment.kt
@@ -1,0 +1,43 @@
+package com.example.playlistmaker.library.ui.fragments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.playlistmaker.R
+import com.example.playlistmaker.databinding.FragmentLibraryBinding
+import com.example.playlistmaker.library.ui.view_model.LibraryViewPageAdapter
+import com.example.playlistmaker.main.ui.fragments.BindingFragments
+import com.google.android.material.tabs.TabLayoutMediator
+
+class LibraryFragment: BindingFragments<FragmentLibraryBinding>() {
+    private var _tabMediator: TabLayoutMediator? = null
+    private val tabMediator get() = _tabMediator!!
+
+    override fun createBinding(
+        inflater: LayoutInflater,
+        container: ViewGroup?
+    ): FragmentLibraryBinding {
+        return FragmentLibraryBinding.inflate(inflater, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        binding.viewPager.adapter = LibraryViewPageAdapter(
+            fragmentManager = childFragmentManager,
+            lifecycle = lifecycle
+        )
+        _tabMediator = TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
+            when(position) {
+                0 -> tab.text = getString(R.string.favorite_tab)
+                1 -> tab.text = getString(R.string.playlist_tab)
+            }
+        }
+        tabMediator.attach()
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        tabMediator.detach()
+    }
+}
