+++ app/src/main/java/com/example/playlistmaker/main/ui/activity/MainActivity.kt
@@ -1,40 +1,38 @@
 package com.example.playlistmaker.main.ui.activity
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.View
-import android.widget.Button
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.fragment.findNavController
+import androidx.navigation.ui.setupWithNavController
 import com.example.playlistmaker.R
-import com.example.playlistmaker.library.ui.activity.LibraryActivity
-import com.example.playlistmaker.search.ui.activity.SearchActivity
-import com.example.playlistmaker.settings.ui.activity.SettingsActivity
+import com.example.playlistmaker.databinding.ActivityMainBinding
+import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class MainActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityMainBinding
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContentView(R.layout.activity_main)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
 
-        val searchBtn = findViewById<Button>(R.id.search_btn)
-        val libraryBtn = findViewById<Button>(R.id.library_btn)
-        val settingsBtn = findViewById<Button>(R.id.settings_btn)
-        val searchClickListener: View.OnClickListener = object : View.OnClickListener {
-            override fun onClick(v: View?) {
-                val searchIntent = Intent(this@MainActivity, SearchActivity::class.java)
-                startActivity(searchIntent)
+        val navHostController = supportFragmentManager.findFragmentById(R.id.mainContainerView)
+        val navController = navHostController?.findNavController()
+        val bottomNavigationView = findViewById<BottomNavigationView>(R.id.bottomNavigationView)
+
+        navController?.let {
+            bottomNavigationView.setupWithNavController(it)
+            it.addOnDestinationChangedListener{ _, destination, _ ->
+                when(destination.id) {
+                    R.id.playerActivity -> {
+                        bottomNavigationView.visibility = View.GONE
+                    }
+                    else -> {
+                        bottomNavigationView.visibility = View.VISIBLE
+                    }
+                }
             }
-        }
-        searchBtn.setOnClickListener(searchClickListener)
-
-        libraryBtn.setOnClickListener {
-            val libraryIntent = Intent(this, LibraryActivity::class.java)
-            startActivity(libraryIntent)
-        }
-        settingsBtn.setOnClickListener {
-            val displayIntent = Intent(this, SettingsActivity::class.java)
-            startActivity(displayIntent)
         }
     }
