+++ app/src/main/res/color/button_nav.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:color="#3772E7"
+        android:state_checked="true" />
+    <item android:color="#1A1B22"/>
+</selector>
