+++ app/src/main/res/layout/fragment_library.xml
@@ -1,11 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
     style="@style/SettingsWindowStyle"
-    android:background="@color/color_background_other">
+    android:background="@color/color_background_other"
+    tools:context=".library.ui.fragments.LibraryFragment"
+    >
 
     <TextView
         android:id="@+id/home"
