+++ gradle/libs.versions.toml
@@ -1,19 +1,23 @@
 [versions]
-agp = "8.4.1"
-kotlin = "1.9.0"
-coreKtx = "1.10.1"
+agp = "8.8.2"
+kotlin = "1.9.24"
+coreKtx = "1.15.0"
 junit = "4.13.2"
-junitVersion = "1.1.5"
-espressoCore = "3.5.1"
-appcompat = "1.6.1"
-material = "1.10.0"
-activity = "1.8.2"
-constraintlayout = "2.1.4"
+junitVersion = "1.2.1"
+espressoCore = "3.6.1"
+appcompat = "1.7.0"
+material = "1.12.0"
+activity = "1.10.1"
+constraintlayout = "2.2.1"
 glide = "4.16.0"
 retrofit = "2.9.0"
 converter-gson = "2.9.0"
 gson = "2.10.1"
 koin = "3.3.0"
+navigation = "2.8.8"
+fragment = "1.8.6"
+viewpager2 = "1.1.0"
+
 
 [libraries]
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
@@ -29,6 +33,10 @@ retrofit = { module = "com.squareup.retrofit2:retrofit", version.ref = "retrofit
 convertergson = { module = "com.squareup.retrofit2:converter-gson", version.ref = "converter-gson" }
 gson = { module = "com.google.code.gson:gson", version.ref = "gson" }
 koin = { group = "io.insert-koin", name = "koin-android", version.ref = "koin" }
+androidx-fragment = { group = "androidx.fragment", name = "fragment-ktx", version.ref = "fragment" }
+androidx-vievpager2 = { group = "androidx.viewpager2", name = "viewpager2", version.ref = "viewpager2" }
+androidx-nav-fragment = { group = "androidx.navigation", name = "navigation-fragment-ktx", version.ref = "navigation" }
+androidx-nav-ui = { group = "androidx.navigation", name = "navigation-ui-ktx", version.ref = "navigation" }
 
 [plugins]
 androidApplication = { id = "com.android.application", version.ref = "agp" }
